+++ services/audit-service/pom.xml
@@ -11,7 +11,7 @@
     <version>1.7-SNAPSHOT</version>
     <description>DATAWAVE Auditing Microservice</description>
     <properties>
-        <version.hadoop>2.6.0-cdh5.9.1</version.hadoop>
+        <version.hadoop>2.6.0</version.hadoop>
         <version.microservice.audit-api>1.4</version.microservice.audit-api>
         <version.microservice.hazelcast-client>1.5-SNAPSHOT</version.microservice.hazelcast-client>
         <version.microservice.starter>1.3</version.microservice.starter>
@@ -125,13 +125,6 @@
         </dependency>
     </dependencies>
     <repositories>
-        <repository>
-            <snapshots>
-                <enabled>false</enabled>
-            </snapshots>
-            <id>cloudera</id>
-            <url>https://repository.cloudera.com/artifactory/cloudera-repos/</url>
-        </repository>
         <repository>
             <!--
                 This repository is a temporary work-around until we can public artifacts to maven central.
