+++ services/audit-service/src/main/java/datawave/microservice/audit/replay/config/ReplayConfig.java
@@ -2,9 +2,9 @@ package datawave.microservice.audit.replay.config;
 
 import com.hazelcast.spring.cache.HazelcastCacheManager;
 import datawave.microservice.audit.replay.status.StatusCache;
-import datawave.microservice.audit.replay.util.LockableCacheInspector;
-import datawave.microservice.audit.replay.util.UniversalLockableCacheInspector;
-import datawave.microservice.audit.replay.util.LockableHazelcastCacheInspector;
+import datawave.microservice.cached.LockableCacheInspector;
+import datawave.microservice.cached.UniversalLockableCacheInspector;
+import datawave.microservice.cached.LockableHazelcastCacheInspector;
 import datawave.microservice.cached.CacheInspector;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
