+++ services/sample_configuration/authorization-dev.yml.example
@@ -5,6 +5,6 @@ server:
   port: 8643
 
 # Don't use a Hazelcast client, but rather use an embedded Hazelcast member. This is
-# only useful for development, and in production, this property shouldnot be set (or
+# only useful for development, and in production, this property should not be set (or
 # should only be set to true)
 hazelcast.client.enabled: ${hzClient:false}
